Unified X11/Wayland code path using _glfwUpdateTaskbarProgressDBusPOSIX()
